Change project bundle id to `io.github.asseco.pst` to be according with the project's repo host domain
