--- src/main/groovy/com/pst/asseco/channels/devops/infrastructure/crawlers/Observable.groovy
@@ -1,8 +1,0 @@
-package com.pst.asseco.channels.devops.infrastructure.crawlers
-
-interface Observable {
-
-    void attach(Observer aObserver)
-
-    void _notify()
-}
