--- src/main/groovy/com/pst/asseco/channels/devops/infrastructure/crawlers/Observer.groovy
@@ -1,6 +1,0 @@
-package com.pst.asseco.channels.devops.infrastructure.crawlers
-
-interface Observer {
-
-    void update()
-}
