+++ src/main/groovy/io/github/asseco/pst/infrastructure/crawlers/FileStorer.groovy
@@ -1,0 +1,46 @@
+package io.github.asseco.pst.infrastructure.crawlers
+
+import io.github.asseco.pst.infrastructure.Project
+import io.github.asseco.pst.infrastructure.utils.Console
+
+/**
+ * This thread is responsible to store, on a given location,
+ * the requirements file of a given Project
+ *
+ * @throws Exception in case it's not able to save the required file
+ */
+class FileStorer extends Thread {
+
+    File workspaceFolder
+    Project project
+
+    FileStorer(Project aProject, File aWorkspaceFolder) {
+        project = aProject
+        workspaceFolder = aWorkspaceFolder
+    }
+
+    @Override
+    void run() {
+
+        storeFile()
+    }
+
+
+    private void storeFile() {
+
+        String filename = "${Project.REQUIREMENTS_FILE}-${project.getVersion().toString()}-${project.versionCommitSha}"
+        Console.print("Storing ${filename} of Project ${project.name}")
+
+        try {
+            File projectFolder = new File(workspaceFolder, project.name)
+            projectFolder.mkdir()
+
+            File requirements = new File(projectFolder, filename)
+            requirements.write(project.requirementsFileContent)
+
+            Console.print("File '${filename}' from '${project.name}' Project successfully stored into ${requirements.getPath()}")
+        } catch (e) {
+            throw new Exception("Unable to store requirements file for Project '${project.name}'. Cause: ${e}")
+        }
+    }
+}
