+++ src/main/groovy/io/github/asseco/pst/infrastructure/crawlers/Observable.groovy
@@ -1,0 +1,8 @@
+package io.github.asseco.pst.infrastructure.crawlers
+
+interface Observable {
+
+    void attach(Observer aObserver)
+
+    void _notify()
+}
