+++ src/main/groovy/io/github/asseco/pst/infrastructure/crawlers/Observer.groovy
@@ -1,0 +1,6 @@
+package io.github.asseco.pst.infrastructure.crawlers
+
+interface Observer {
+
+    void update()
+}
